Update AdContainer.js. Renamed SeacrhedAds.js -> SearchedAds.js.
